Reorganize farms feature
