+++ v2/styles/tailwind.css
@@ -1,9 +1,3 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-@layer utilities {
-  .border-spacing-2 {
-    border-spacing: 0 8px;
-  }
-}
