+++ v2/types/index.ts
@@ -3,3 +3,12 @@ import { FC } from "react";
 export interface PickleFinancePage extends FC {
   PageTitle: FC;
 }
+
+export interface Farm {
+  asset: string;
+  iconSrc: string;
+  earned: number;
+  deposited: number;
+  apy: string;
+  liquidity: number;
+}
